tumour_finder/app/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html 
torch==2.1.2+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html 
torchvision==0.16.2+cpu
Pillow==10.1.0
annotated-types==0.5.0
anyio==3.7.1
click==8.1.5
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
pydantic==2.0.2
pydantic-core==2.1.2
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0

